requirements.txt
google-generativeai>=0.8.0
fastapi
uvicorn
python-dotenv
mcp
fastmcp
google-cloud-aiplatform
pydantic>=2.0
